apps/server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anima-server"
version = "0.1.0"
description = "Python backend for ANIMA."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.5",
  "argon2-cffi>=25.1.0",
  "cryptography>=45.0.0",
  "fastapi>=0.115.12",
  "httpx>=0.28.1",
  "jinja2>=3.1.6",
  "pydantic-settings>=2.11.0",
  "sqlalchemy>=2.0.43",
  "sqlcipher3>=0.5.4",
  "uvicorn[standard]>=0.35.0",
  "chromadb>=1.0.0"
]

[dependency-groups]
postgres = ["psycopg[binary]>=3.2.9"]
dev = [
  "pytest>=8.4.0",
  "pytest-asyncio>=1.3.0",
  "ruff>=0.12.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/anima_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
